Stop crouching being activated in vehicle
Improved the check for if ped is in vehicle.  Stopped crouch from being activated causing the player to be in crouched state when exiting vehicle.  This is just a small tweak to this script to assist with the issue where when toggling the crouch state afterwards would cause player to be standing but also not be able to go into the first person camera.